chore(localized-text-input): move to packages
