+++ packages/components/inputs/localized-text-input/src/index.js
@@ -1,1 +1,3 @@
 export { default } from './localized-text-input';
+
+export { default as version } from './version';
