+++ packages/components/inputs/localized-text-input/src/localized-text-input.story.js
@@ -10,8 +10,8 @@ import {
   object,
 } from '@storybook/addon-knobs/react';
 import { ErrorMessage } from '@commercetools-uikit/messages';
-import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
 import LocalizedTextInput from './localized-text-input';
 
 storiesOf('Components|Inputs', module)
