+++ packages/components/inputs/localized-text-input/src/localized-text-input.visualspec.js
