+++ packages/components/inputs/localized-text-input/src/version.js
